Используем float64

+++ adapters/common.go
@@ -1,11 +1,11 @@
 package adapters
 
 type MeasurementSchema struct {
-    Humidity float32
-    Pressure float32
-    Precipitation float32
-    Temp float32
-    Wind float32
+    Humidity float64
+    Pressure float64
+    Precipitation float64
+    Temp float64
+    Wind float64
 }
 
 type MeasurementArray []MeasurementSchema
